fix: refetch local graphql schema when outdated
