+++ tests/graphql/test_graphql_client.py
@@ -1,3 +1,11 @@
+import os
+from pathlib import Path
+from unittest import mock
+
+import graphql
+import pytest
+
+from kili.exceptions import GraphQLError
 from kili.graphql.graphql_client import GraphQLClient, GraphQLClientName
 
 
@@ -13,3 +21,64 @@ def test_graphql_client_cache_cant_get_kili_version(mocker):
         client_name=GraphQLClientName.SDK,
         verify=True,
     )
+
+
+@pytest.mark.parametrize(
+    "query",
+    [
+        "query MyQuery { my_query_is_clearly_not_valid_according_to_the_kili_schema { id } }",
+        "query { projects { this_field_does_not_exist } }",
+    ],
+)
+def test_gql_bad_query_local_validation(query):
+    """test gql validation against local schema"""
+    api_endpoint = os.getenv(
+        "KILI_API_ENDPOINT", "https://cloud.kili-technology.com/api/label/v2/graphql"
+    )
+    api_key = os.getenv("KILI_API_KEY")
+
+    client = GraphQLClient(
+        endpoint=api_endpoint,  # type: ignore
+        api_key=api_key,  # type: ignore
+        client_name=GraphQLClientName.SDK,
+        verify=True,
+    )
+
+    with pytest.raises(GraphQLError) as exc_info:
+        client.execute(query)
+
+    assert isinstance(exc_info.value.__cause__, graphql.GraphQLError)
+
+
+def test_graphql_client_cache(mocker):
+    SCHEMA_PATH = Path.home() / ".cache" / "kili" / "graphql" / "schema.graphql"
+    mocker.patch.object(GraphQLClient, "_get_graphql_schema_path", return_value=SCHEMA_PATH)
+
+    api_endpoint = os.getenv(
+        "KILI_API_ENDPOINT", "https://cloud.kili-technology.com/api/label/v2/graphql"
+    )
+    api_key = os.getenv("KILI_API_KEY")
+
+    if SCHEMA_PATH.is_file():
+        SCHEMA_PATH.unlink()
+
+    _ = GraphQLClient(
+        endpoint=api_endpoint,  # type: ignore
+        api_key=api_key,  # type: ignore
+        client_name=GraphQLClientName.SDK,
+        verify=True,
+    )
+    assert SCHEMA_PATH.is_file()  # schema cached
+    assert SCHEMA_PATH.stat().st_size > 0  # schema not empty
+
+    with mock.patch("kili.graphql.graphql_client.print_schema") as mocked_print_schema:
+        _ = GraphQLClient(
+            endpoint=api_endpoint,  # type: ignore
+            api_key=api_key,  # type: ignore
+            client_name=GraphQLClientName.SDK,
+            verify=True,
+        )
+        mocked_print_schema.assert_not_called()
+
+    if SCHEMA_PATH.is_file():
+        SCHEMA_PATH.unlink()
